- Refactored ilm-related manager funcs into their own files (it was getting too big) - Added missing mock functions to MockManager so that make test builds correctly.
