+++ controller/manager/manager.go
@@ -10,17 +10,13 @@ import (
 	r "github.com/dancannon/gorethink"
 	"github.com/gorilla/sessions"
 	"github.com/samalba/dockerclient"
-	c "github.com/shipyard/shipyard/checker"
-	apiClient "github.com/shipyard/shipyard/client"
 	"github.com/shipyard/shipyard/model"
 	"github.com/shipyard/shipyard/model/dockerhub"
 	"github.com/shipyard/shipyard/utils/auth"
 	"github.com/shipyard/shipyard/version"
 	"net"
 	"net/http"
-	"regexp"
 	"strings"
-	"sync"
 	"time"
 )
 
@@ -781,1158 +777,6 @@ func (m DefaultManager) Node(name string) (*model.Node, error) {
 
 	return nil, nil
 }
-
-// methods related to the Project structure
-func (m DefaultManager) Projects() ([]*model.Project, error) {
-	// TODO: consider making sorting customizable
-	// TODO: should filter by authorization
-	// Return all projects **WITHOUT** their images embedded
-	res, err := r.Table(tblNameProjects).OrderBy(r.Asc("creationTime")).Run(m.session)
-	if err != nil {
-		return nil, err
-	}
-	projects := []*model.Project{}
-	if err := res.All(&projects); err != nil {
-		return nil, err
-	}
-
-	return projects, nil
-}
-
-func (m DefaultManager) Project(id string) (*model.Project, error) {
-
-	var project *model.Project
-
-	res, err := r.Table(tblNameProjects).Filter(map[string]string{"id": id}).Run(m.session)
-
-	if err != nil {
-		return nil, err
-	}
-	if res.IsNil() {
-		return nil, ErrProjectDoesNotExist
-	}
-	if err := res.One(&project); err != nil {
-		return nil, err
-	}
-
-	project.Images, err = m.GetImages(project.ID)
-
-	if err != nil {
-		return nil, ErrProjectImagesProblem
-	}
-	project.Tests, err = m.GetTests(project.ID)
-
-	if err != nil {
-		return nil, ErrProjectTestsProblem
-	}
-
-	return project, nil
-}
-
-func (m DefaultManager) SaveProject(project *model.Project) error {
-	var eventType string
-	proj, err := m.Project(project.ID)
-
-	if err != nil && err != ErrProjectDoesNotExist {
-		return err
-	}
-	if proj != nil {
-		return ErrProjectExists
-	}
-	project.CreationTime = time.Now().UTC()
-	project.UpdateTime = project.CreationTime
-	// TODO: find a way to retrieve the current user
-	project.Author = "author"
-
-	//create the project
-	response, err := r.Table(tblNameProjects).Insert(project).RunWrite(m.session)
-
-	if err != nil {
-		return err
-	}
-
-	// rethinkDB returns the ID as the first element of the GeneratedKeys slice
-	// TODO: this method seems brittle, should contact the gorethink dev team for insight on this.
-	project.ID = func() string {
-		if len(response.GeneratedKeys) > 0 {
-			return string(response.GeneratedKeys[0])
-		}
-		return ""
-	}()
-
-	//add the project ID to the images and save them in the Images table
-	// TODO: investigate how to do a bulk insert
-	for _, img := range project.Images {
-		img.ProjectId = project.ID
-		response, err = r.Table(tblNameImages).Insert(img).RunWrite(m.session)
-
-		if err != nil {
-			return err
-		}
-	}
-	//add project ID to the tests and save them in the Tests table
-	// TODO: investigate how to do a bulk insert
-	for _, test := range project.Tests {
-		test.ProjectId = project.ID
-		response, err = r.Table(tblNameTests).Insert(test).RunWrite(m.session)
-
-		if err != nil {
-			return err
-		}
-	}
-
-	eventType = "add-project"
-	m.logEvent(eventType, fmt.Sprintf("id=%s, name=%s", project.ID, project.Name), []string{"security"})
-	return nil
-}
-
-func (m DefaultManager) UpdateProject(project *model.Project) error {
-	var eventType string
-	// check if exists; if so, update
-	proj, err := m.Project(project.ID)
-	if err != nil && err != ErrProjectDoesNotExist {
-		return err
-	}
-	// update
-	if proj != nil {
-		updates := map[string]interface{}{
-			"name":        project.Name,
-			"description": project.Description,
-			"status":      project.Status,
-			"needsBuild":  project.NeedsBuild,
-			"updateTime":  time.Now().UTC(),
-			// TODO: find a way to retrieve the current user
-			"updatedBy": "updater",
-		}
-
-		//TODO: Find a more elegant approach
-		// Retrieve images by projectId and delete them by their primary key id generated by rethink
-		res, err := r.Table(tblNameImages).Filter(map[string]string{"projectId": proj.ID}).Run(m.session)
-		if err != nil {
-			return err
-		}
-		oldImages := []*model.Image{}
-		if err := res.All(&oldImages); err != nil {
-			return err
-		}
-
-		// Remove existing images for this project
-		for _, oldImage := range oldImages {
-			if _, err := r.Table(tblNameImages).Filter(map[string]string{"id": oldImage.ID}).Delete().Run(m.session); err != nil {
-				return err
-			}
-		}
-
-		// Insert all the images that are incoming from the request which should have the new and old ones
-		// TODO: investigate how we can do bulk insert
-		for _, newImage := range project.Images {
-			newImage.ProjectId = proj.ID
-			if _, err := r.Table(tblNameImages).Insert(newImage).RunWrite(m.session); err != nil {
-				return err
-			}
-		}
-
-		//TODO: Find a more elegant approach
-		// Retrieve tests by projectId and delete them by their primary key id generated by rethink
-		res, err = r.Table(tblNameTests).Filter(map[string]string{"projectId": proj.ID}).Run(m.session)
-		if err != nil {
-			return err
-		}
-		oldTests := []*model.Test{}
-		if err := res.All(&oldTests); err != nil {
-			return err
-		}
-
-		// Remove existing tests for this project
-		for _, oldTest := range oldTests {
-			if _, err := r.Table(tblNameTests).Filter(map[string]string{"id": oldTest.ID}).Delete().Run(m.session); err != nil {
-				return err
-			}
-		}
-
-		// Insert all the tests that are incoming from the request which should have the new and old ones
-		// TODO: investigate how we can do bulk insert
-		for _, newTest := range project.Tests {
-			newTest.ProjectId = proj.ID
-			if _, err := r.Table(tblNameTests).Insert(newTest).RunWrite(m.session); err != nil {
-				return err
-			}
-		}
-		if _, err := r.Table(tblNameProjects).Filter(map[string]string{"id": project.ID}).Update(updates).RunWrite(m.session); err != nil {
-			return err
-		}
-
-		eventType = "update-project"
-	}
-
-	m.logEvent(eventType, fmt.Sprintf("id=%s, name=%s", project.ID, project.Name), []string{"security"})
-
-	return nil
-}
-
-func (m DefaultManager) DeleteProject(project *model.Project) error {
-	res, err := r.Table(tblNameProjects).Filter(map[string]string{"id": project.ID}).Delete().Run(m.session)
-	if err != nil {
-		return err
-	}
-
-	if res.IsNil() {
-		return ErrProjectDoesNotExist
-	}
-	res, err = r.Table(tblNameImages).Filter(map[string]string{"projectId": project.ID}).Run(m.session)
-	if err != nil {
-		return err
-	}
-	imagesToDelete := []*model.Image{}
-	if err := res.All(&imagesToDelete); err != nil {
-		return err
-	}
-
-	// Remove existing images for this project
-	for _, imgToDelete := range imagesToDelete {
-		if _, err := r.Table(tblNameImages).Filter(map[string]string{"id": imgToDelete.ID}).Delete().Run(m.session); err != nil {
-			return err
-		}
-	}
-
-	res, err = r.Table(tblNameTests).Filter(map[string]string{"projectId": project.ID}).Run(m.session)
-	if err != nil {
-		return err
-	}
-	testsToDelete := []*model.Test{}
-	if err := res.All(&testsToDelete); err != nil {
-		return err
-	}
-
-	// Remove existing tests for this project
-	for _, testToDelete := range testsToDelete {
-		if _, err := r.Table(tblNameTests).Filter(map[string]string{"id": testToDelete.ID}).Delete().Run(m.session); err != nil {
-			return err
-		}
-	}
-
-	m.logEvent("delete-project", fmt.Sprintf("id=%s, name=%s", project.ID, project.Name), []string{"security"})
-
-	return nil
-}
-
-func (m DefaultManager) DeleteAllProjects() error {
-	_, err := r.Table(tblNameProjects).Delete().Run(m.session)
-
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// end methods related to the project structure
-
-// check if an image exists
-func (m DefaultManager) VerifyIfImageExistsLocally(imageToCheck string) bool {
-
-	//images, err := m.client.ListImages(true)
-
-	images, err := apiClient.GetLocalImages(m.DockerClient().URL.String())
-
-	if err != nil {
-		log.Error(err)
-		return false
-	}
-
-	for _, img := range images {
-		imageRepoTags := img.RepoTags
-		for _, imageRepoTag := range imageRepoTags {
-			if imageRepoTag == imageToCheck {
-				fmt.Printf("Image %s exists locally as %s \n", imageToCheck, imageRepoTag)
-				return true
-			}
-		}
-	}
-
-	return false
-}
-
-func (m DefaultManager) PullImage(imageNameAndTag string) error {
-	auth := dockerclient.AuthConfig{"", "", ""}
-
-	fmt.Printf("Image does not exist locally. Pulling image %s ... \n", imageNameAndTag)
-	//get registry
-	match, _ := regexp.MatchString(":[0-9]{4}/", imageNameAndTag)
-	if match {
-		parts := strings.Split(imageNameAndTag, "/")
-		address := "https://" + parts[0]
-
-		registry, err := m.RegistryByAddress(address)
-		auth = dockerclient.AuthConfig{registry.Username, registry.Password, ""}
-		if err != nil {
-			log.Error(err)
-			return err
-		}
-	}
-	ticker := time.NewTicker(time.Second * 15)
-	go func() {
-		for t := range ticker.C {
-			fmt.Print("Time: ", t.UTC())
-			fmt.Printf(" Pulling image: %s. Please be patient while the process finishes ... \n", imageNameAndTag)
-		}
-	}()
-	err2 := m.client.PullImage(imageNameAndTag, &auth)
-
-	if err2 != nil {
-		fmt.Printf("Could not pull image %s ... \n %s \n", imageNameAndTag, err2)
-		return err2
-		ticker.Stop()
-	}
-	ticker.Stop()
-
-	return nil
-}
-
-//methods related to the Image structure
-func (m DefaultManager) GetImages(projectId string) ([]*model.Image, error) {
-
-	res, err := r.Table(tblNameImages).Filter(map[string]string{"projectId": projectId}).Run(m.session)
-	if err != nil {
-		return nil, err
-	}
-	images := []*model.Image{}
-	if err := res.All(&images); err != nil {
-		return nil, err
-	}
-	return images, nil
-}
-
-func (m DefaultManager) GetImage(projectId string, imageId string) (*model.Image, error) {
-	var image *model.Image
-	res, err := r.Table(tblNameImages).Filter(map[string]string{"id": imageId}).Run(m.session)
-	if err != nil {
-		return nil, err
-	}
-	if res.IsNil() {
-		return nil, ErrImageDoesNotExist
-	}
-	if err := res.One(&image); err != nil {
-		return nil, err
-	}
-
-	return image, nil
-}
-
-func (m DefaultManager) CreateImage(projectId string, image *model.Image) error {
-	var eventType string
-	image.ProjectId = projectId
-	response, err := r.Table(tblNameImages).Insert(image).RunWrite(m.session)
-	if err != nil {
-
-		return err
-	}
-	image.ID = func() string {
-		if len(response.GeneratedKeys) > 0 {
-			return string(response.GeneratedKeys[0])
-		}
-		return ""
-	}()
-	eventType = "add-image"
-
-	m.logEvent(eventType, fmt.Sprintf("id=%s", image.ID), []string{"security"})
-	return nil
-}
-
-func (m DefaultManager) UpdateImage(projectId string, image *model.Image) error {
-	var eventType string
-	// check if exists; if so, update
-	rez, err := m.GetImage(projectId, image.ID)
-	if err != nil && err != ErrImageDoesNotExist {
-		return err
-	}
-	// update
-	if rez != nil {
-		updates := map[string]interface{}{
-			"name":           image.Name,
-			"imageId":        image.ImageId,
-			"tag":            image.Tag,
-			"description":    image.Description,
-			"location":       image.Location,
-			"skipImageBuild": image.SkipImageBuild,
-			"projectId":      image.ProjectId,
-		}
-		if _, err := r.Table(tblNameImages).Filter(map[string]string{"id": image.ID}).Update(updates).RunWrite(m.session); err != nil {
-			return err
-		}
-
-		eventType = "update-image"
-	}
-
-	m.logEvent(eventType, fmt.Sprintf("id=%s", image.ID), []string{"security"})
-	return nil
-}
-
-func (m DefaultManager) DeleteImage(projectId string, imageId string) error {
-	res, err := r.Table(tblNameImages).Filter(map[string]string{"id": imageId}).Delete().Run(m.session)
-	if err != nil {
-		return err
-	}
-
-	if res.IsNil() {
-		return ErrImageDoesNotExist
-	}
-
-	m.logEvent("delete-image", fmt.Sprintf("id=%s", imageId), []string{"security"})
-
-	return nil
-}
-
-func (m DefaultManager) DeleteAllImages() error {
-	_, err := r.Table(tblNameImages).Delete().Run(m.session)
-
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-//methods related to Test structure
-
-func (m DefaultManager) GetTests(projectId string) ([]*model.Test, error) {
-
-	res, err := r.Table(tblNameTests).Filter(map[string]string{"projectId": projectId}).Run(m.session)
-	if err != nil {
-		return nil, err
-	}
-	tests := []*model.Test{}
-	if err := res.All(&tests); err != nil {
-		return nil, err
-	}
-	return tests, nil
-}
-
-func (m DefaultManager) GetTest(projectId, testId string) (*model.Test, error) {
-	var test *model.Test
-	res, err := r.Table(tblNameTests).Filter(map[string]string{"id": testId}).Run(m.session)
-	if err != nil {
-		return nil, err
-	}
-	if res.IsNil() {
-		return nil, ErrTestDoesNotExist
-	}
-	if err := res.One(&test); err != nil {
-		return nil, err
-	}
-
-	return test, nil
-}
-
-func (m DefaultManager) CreateTest(projectId string, test *model.Test) error {
-	var eventType string
-	test.ProjectId = projectId
-	response, err := r.Table(tblNameTests).Insert(test).RunWrite(m.session)
-	if err != nil {
-
-		return err
-	}
-	test.ID = func() string {
-		if len(response.GeneratedKeys) > 0 {
-			return string(response.GeneratedKeys[0])
-		}
-		return ""
-	}()
-	eventType = "add-test"
-
-	m.logEvent(eventType, fmt.Sprintf("id=%s", test.ID), []string{"security"})
-	return nil
-}
-
-func (m DefaultManager) UpdateTest(projectId string, test *model.Test) error {
-	var eventType string
-	// check if exists; if so, update
-	rez, err := m.GetTest(projectId, test.ID)
-	if err != nil && err != ErrTestDoesNotExist {
-		return err
-	}
-	// update
-	if rez != nil {
-		updates := map[string]interface{}{
-			"projectId":        test.ProjectId,
-			"description":      test.Description,
-			"name":             test.Name,
-			"targets":          test.Targets,
-			"selectedTestType": test.SelectedTestType,
-			"ProviderType":     test.Provider.ProviderType,
-			"providerName":     test.Provider.ProviderName,
-			"providerTest":     test.Provider.ProviderTest,
-			"onSuccess":        test.Tagging.OnSuccess,
-			"onFailure":        test.Tagging.OnFailure,
-			"fromTag":          test.FromTag,
-			"parameters":       test.Parameters,
-		}
-		if _, err := r.Table(tblNameTests).Filter(map[string]string{"id": test.ID}).Update(updates).RunWrite(m.session); err != nil {
-			return err
-		}
-
-		eventType = "update-test"
-	}
-
-	m.logEvent(eventType, fmt.Sprintf("id=%s", test.ID), []string{"security"})
-	return nil
-}
-
-func (m DefaultManager) DeleteTest(projectId string, testId string) error {
-	res, err := r.Table(tblNameTests).Filter(map[string]string{"id": testId}).Delete().Run(m.session)
-	if err != nil {
-		return err
-	}
-
-	if res.IsNil() {
-		return ErrTestDoesNotExist
-	}
-
-	m.logEvent("delete-test", fmt.Sprintf("id=%s", testId), []string{"security"})
-	return nil
-}
-func (m DefaultManager) DeleteAllTests() error {
-	_, err := r.Table(tblNameTests).Delete().Run(m.session)
-
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-//methods related to the Build structure
-func (m DefaultManager) GetBuilds(projectId string, testId string) ([]*model.Build, error) {
-	res, err := r.Table(tblNameBuilds).Filter(map[string]string{"projectId": projectId, "testId": testId}).Run(m.session)
-	if err != nil {
-		return nil, err
-	}
-	builds := []*model.Build{}
-	if err := res.All(&builds); err != nil {
-		return nil, err
-	}
-	return builds, nil
-}
-
-func (m DefaultManager) GetBuild(projectId string, testId string, buildId string) (*model.Build, error) {
-	return m.GetBuildById(buildId)
-}
-
-func (m DefaultManager) GetBuildById(buildId string) (*model.Build, error) {
-	res, err := r.Table(tblNameBuilds).Filter(map[string]string{"id": buildId}).Run(m.session)
-	if err != nil {
-		return nil, err
-	}
-	if res.IsNil() {
-		return nil, ErrBuildDoesNotExist
-	}
-	var build *model.Build
-	if err := res.One(&build); err != nil {
-		return nil, err
-	}
-	return build, nil
-}
-
-func (m DefaultManager) GetBuildStatus(projectId string, testId string, buildId string) (string, error) {
-	build, err := m.GetBuildById(buildId)
-
-	if err != nil {
-		log.Errorf("Could not get build status for build= %s err= %s", buildId, err.Error())
-		return "", err
-	}
-
-	if build.Status == nil {
-		errMsg := fmt.Sprintf("Could not get build status for build= %s, status is nil", buildId, err.Error())
-		log.Errorf(errMsg)
-		return "", errors.New(errMsg)
-	}
-	return build.Status.Status, nil
-}
-func (m DefaultManager) GetBuildResults(projectId string, testId string, buildId string) ([]*model.BuildResult, error) {
-	build, err := m.GetBuildById(buildId)
-
-	if err != nil {
-		log.Errorf("Could not get results for build= %s err= %", buildId, err.Error())
-		return nil, err
-	}
-
-	return build.Results, nil
-}
-func (m DefaultManager) CreateBuild(projectId string, testId string, buildAction *model.BuildAction) (string, error) {
-
-	var eventType string
-	eventType = eventType
-	var build *model.Build
-
-	// In order to create a build we should get a start action
-	if buildAction.Action != model.BuildStartActionLabel {
-		log.Errorf("Build action should be %s, but received %s, error = %s",
-			model.BuildStartActionLabel,
-			buildAction.Action,
-			ErrBuildActionNotSupported.Error(),
-		)
-		return "", ErrBuildActionNotSupported
-	}
-
-	// Instantiate a new Build object and fill out some fields
-	build = &model.Build{}
-	build.TestId = testId
-	build.ProjectId = projectId
-	build.StartTime = time.Now()
-
-	// we change the build's buildStatus to submitted
-	build.Status = &model.BuildStatus{Status: model.BuildStatusNewLabel}
-
-	// Get the project related to the Test / Build
-	project, err := m.Project(projectId)
-	if err != nil && err != ErrProjectDoesNotExist {
-		return "", err
-	}
-
-	// Get the Test and its TargetArtifacts
-	test, err := m.GetTest(projectId, testId)
-	if err != nil && err != ErrTestDoesNotExist {
-		return "", err
-	}
-	targetArtifacts := test.Targets
-
-	// we get the ids for the targets we want to test
-	targetIds := []string{}
-	for _, target := range targetArtifacts {
-		targetIds = append(targetIds, target.ID)
-
-	}
-	// Retrieve the images from the projectId
-	// TODO: Investigate if we can query db for the images matching the Ids in the TargetArtifacts
-	projectImages, err := m.GetImages(projectId)
-	if err != nil && err != ErrProjectImagesProblem {
-		return "", err
-	}
-
-	// Collect the images that are TargetArtifacts
-	// by comparing the ImageID with the ArtifactId
-	imagesToBuild := []*model.Image{}
-	for _, image := range projectImages {
-		for _, artifactId := range targetIds {
-			if image.ID == artifactId {
-				imagesToBuild = append(imagesToBuild, image)
-			}
-		}
-	}
-
-	// Store the Build in the table in rethink db
-	response, err := r.Table(tblNameBuilds).Insert(build).RunWrite(m.session)
-	if err != nil {
-		return "", err
-	}
-
-	build.ID = func() string {
-		if len(response.GeneratedKeys) > 0 {
-			return string(response.GeneratedKeys[0])
-		}
-		return ""
-	}()
-
-	// Start a goroutine that will execute the build non-blocking
-	// TODO: this go routine should be replaced eventually to a call to the provider bridge / engine
-	go func() {
-
-		var wg sync.WaitGroup
-		log.Printf("Processing %d image(s)", len(imagesToBuild))
-		// For each image that we target in the test, try to run a build / verification
-		for _, image := range imagesToBuild {
-			name := image.Name + ":" + image.Tag
-			log.Printf("Processing image=%s", name)
-			wg.Add(1)
-
-			// Run the verification concurrently for each image and then block to wait for all to finish.
-			go m.executeBuildTask(
-				project,
-				test,
-				build,
-				name,
-				image,
-				&wg,
-			)
-		}
-		// Block the outer goroutine until ALL the inner goroutines finish
-		wg.Wait()
-	}()
-
-	// TODO: all these event types should be refactored as constants
-	eventType = "add-build"
-	m.logEvent(eventType, fmt.Sprintf("id=%s", build.ID), []string{"security"})
-	return build.ID, nil
-}
-
-// Executes a BuilTask in the background as part of a wait group.
-func (m DefaultManager) executeBuildTask(
-	project *model.Project,
-	test *model.Test,
-	build *model.Build,
-	name string,
-	image *model.Image,
-	wg *sync.WaitGroup,
-) {
-	// TODO: need to revisit API spec, there are just too many redundant "Result" types stored,
-	// TODO: these should probably just be views of the BuildResults
-	// TODO: need to set the author to the real user
-	// TODO: use model.NewResult() instead
-	result := &model.Result{
-		BuildId:     build.ID,
-		Author:      "author",
-		ProjectId:   project.ID,
-		Description: project.Description,
-		Updater:     "author",
-		CreateDate:  time.Now(),
-	}
-
-	// TODO: use model.NewTestResult() instead
-	testResult := model.TestResult{}
-	testResult.Date = time.Now()
-	testResult.TestId = test.ID
-	testResult.BuildId = build.ID
-	testResult.TestName = test.Name
-	testResult.ImageName = name
-	testResult.BuildId = build.ID
-
-	thisImageName := name
-
-	// When the goroutine finishes, mark this wait group item as done.
-	// TODO: perhaps do this only if wg != nil?
-	defer wg.Done()
-
-	// Check to see if the image exists locally, if not, try to pull it.
-	if !m.VerifyIfImageExistsLocally(thisImageName) {
-		log.Printf("Image %s not available locally, will try to pull...", thisImageName)
-		if err := m.PullImage(thisImageName); err != nil {
-			log.Errorf("Error pulling image %s", thisImageName)
-			return
-		}
-	}
-
-	// Get all local images
-	localImages, err := apiClient.GetLocalImages(m.DockerClient().URL.String())
-
-	// get the docker image id and append it to the test results
-	for _, localImage := range localImages {
-		imageRepoTags := localImage.RepoTags
-		for _, imageRepoTag := range imageRepoTags {
-			if imageRepoTag == thisImageName {
-				//image.DockerImageId = localImage.ID
-				testResult.DockerImageId = localImage.ID
-				image.ImageId = localImage.ID
-			}
-		}
-	}
-
-	m.UpdateBuildStatus(build.ID, "running")
-	existingResult, _ := m.GetResults(project.ID)
-
-	// Once the image is available, try to test it with Clair
-	log.Printf("Will attempt to test image %s with Clair...", thisImageName)
-	resultsSlice, isSafe, err := c.CheckImage(image)
-
-	targetArtifact := model.NewTargetArtifact(
-		image.ID,
-		model.TargetArtifactImageType,
-		image,
-	)
-	buildResult := model.NewBuildResult(build.ID, targetArtifact, resultsSlice)
-
-	m.UpdateBuildResults(build.ID, *buildResult)
-	finishLabel := "finished_failed"
-
-	if isSafe && err == nil {
-		// if we don't get an error and we get the isSafe flag == true
-		// we mark the test for the image as successful
-		finishLabel = "finished_success"
-		log.Infof("Image %s is safe!", thisImageName)
-	} else {
-		log.Errorf("Image %s is NOT safe :(", thisImageName)
-	}
-
-	m.UpdateBuildStatus(build.ID, finishLabel)
-
-	testResult.SimpleResult.Status = finishLabel
-	testResult.EndDate = time.Now()
-	testResult.Blocker = false
-	result.TestResults = append(result.TestResults, &testResult)
-	result.LastUpdate = time.Now()
-
-	if existingResult != nil {
-		m.UpdateResult(project.ID, result)
-
-	} else {
-		m.CreateResult(project.ID, result)
-	}
-
-}
-
-func (m DefaultManager) UpdateBuild(projectId string, testId string, buildId string, buildAction *model.BuildAction) error {
-	var eventType string
-
-	// check if exists; if so, update
-	tmpBuild, err := m.GetBuild(projectId, testId, buildId)
-	if err != nil && err != ErrBuildDoesNotExist {
-		return err
-	}
-	// update
-	if tmpBuild != nil {
-		if buildAction.Action == "stop" {
-			tmpBuild.Status.Status = "stopped"
-			tmpBuild.EndTime = time.Now()
-			// go StopCurrentBuildFromClair
-		}
-		if buildAction.Action == "restart" {
-			tmpBuild.Status.Status = "restarted"
-			tmpBuild.EndTime = time.Now()
-			// go RestartCurrentBuildFromClair
-
-		}
-
-		if _, err := r.Table(tblNameBuilds).Filter(map[string]string{"id": buildId}).Update(tmpBuild).RunWrite(m.session); err != nil {
-			return err
-		}
-
-		eventType = "update-build"
-	}
-
-	m.logEvent(eventType, fmt.Sprintf("id=%s", buildId), []string{"security"})
-
-	return nil
-
-}
-
-func (m DefaultManager) UpdateBuildResults(buildId string, result model.BuildResult) error {
-	var eventType string
-	build, err := m.GetBuildById(buildId)
-	if err != nil {
-		return err
-	}
-	build.Results = append(build.Results, &result)
-
-	if _, err := r.Table(tblNameBuilds).Filter(map[string]string{"id": buildId}).Update(build).RunWrite(m.session); err != nil {
-		return err
-	}
-
-	eventType = "update-build-results"
-
-	m.logEvent(eventType, fmt.Sprintf("id=%s", buildId), []string{"security"})
-
-	return nil
-}
-func (m DefaultManager) UpdateBuildStatus(buildId string, status string) error {
-	var eventType string
-	build, err := m.GetBuildById(buildId)
-	if err != nil {
-		return err
-	}
-	build.Status.Status = status
-
-	if _, err := r.Table(tblNameBuilds).Filter(map[string]string{"id": buildId}).Update(build).RunWrite(m.session); err != nil {
-		return err
-	}
-
-	eventType = "update-build-status"
-
-	m.logEvent(eventType, fmt.Sprintf("id=%s", buildId), []string{"security"})
-
-	return nil
-}
-func (m DefaultManager) DeleteBuild(projectId string, testId string, buildId string) error {
-	build, err := r.Table(tblNameBuilds).Filter(map[string]string{"id": buildId}).Delete().Run(m.session)
-	if err != nil {
-		return err
-	}
-
-	if build.IsNil() {
-		return ErrBuildDoesNotExist
-	}
-
-	m.logEvent("delete-build", fmt.Sprintf("id=%s", buildId), []string{"security"})
-
-	return nil
-}
-
-func (m DefaultManager) DeleteAllBuilds() error {
-	_, err := r.Table(tblNameBuilds).Delete().Run(m.session)
-
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// Methods related to the results structure
-func (m DefaultManager) GetResults(projectId string) (*model.Result, error) {
-
-	res, err := r.Table(tblNameResults).Filter(map[string]string{"projectId": projectId}).Run(m.session)
-	if err != nil {
-		return nil, err
-	}
-	var result *model.Result
-	if err := res.One(&result); err != nil {
-		return nil, err
-	}
-	return result, nil
-}
-
-func (m DefaultManager) GetResult(projectId, resultId string) (*model.Result, error) {
-	res, err := r.Table(tblNameResults).Filter(map[string]string{"id": resultId}).Run(m.session)
-	if err != nil {
-		return nil, err
-	}
-	if res.IsNil() {
-		return nil, ErrResultDoesNotExist
-	}
-	var result *model.Result
-	if err := res.One(&result); err != nil {
-		return nil, err
-	}
-	return result, nil
-}
-
-func (m DefaultManager) CreateResult(projectId string, result *model.Result) error {
-	var eventType string
-
-	tmpResult, err := m.GetResult(projectId, result.ID)
-	if err != nil && err != ErrResultDoesNotExist {
-		return err
-	}
-
-	if tmpResult != nil {
-		return ErrResultExists
-	}
-
-	result.ProjectId = projectId
-	response, err := r.Table(tblNameResults).Insert(result).RunWrite(m.session)
-
-	if err != nil {
-		return err
-	}
-	eventType = "add-result"
-
-	result.ID = func() string {
-		if len(response.GeneratedKeys) > 0 {
-			return string(response.GeneratedKeys[0])
-		}
-		return ""
-	}()
-
-	m.logEvent(eventType, fmt.Sprintf("id=%s", result.ID), []string{"security"})
-
-	return nil
-}
-func (m DefaultManager) UpdateResult(projectId string, inputResult *model.Result) error {
-	var eventType string
-
-	// check if exists; if so, update
-	existingResult, err := m.GetResults(projectId)
-	if err != nil && err != ErrResultDoesNotExist {
-		return err
-	}
-	// update
-	if existingResult != nil {
-		for _, result := range inputResult.TestResults {
-			existingResult.TestResults = append(existingResult.TestResults, result)
-		}
-
-		if _, err := r.Table(tblNameResults).Filter(map[string]string{"projectId": projectId}).Update(existingResult).RunWrite(m.session); err != nil {
-			return err
-		}
-
-		eventType = "update-result"
-	}
-
-	m.logEvent(eventType, fmt.Sprintf("id=%s", existingResult.ID), []string{"security"})
-
-	return nil
-}
-func (m DefaultManager) DeleteResult(projectId string, resultId string) error {
-	res, err := r.Table(tblNameResults).Filter(map[string]string{"id": resultId}).Delete().Run(m.session)
-	if err != nil {
-		return err
-	}
-
-	if res.IsNil() {
-		return ErrResultDoesNotExist
-	}
-
-	m.logEvent("delete-result", fmt.Sprintf("id=%s", resultId), []string{"security"})
-
-	return nil
-}
-func (m DefaultManager) DeleteAllResults() error {
-	_, err := r.Table(tblNameResults).Delete().Run(m.session)
-
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// Methods related to the Provider structure
-func (m DefaultManager) GetProviders() ([]*model.Provider, error) {
-
-	res, err := r.Table(tblNameProviders).OrderBy(r.Asc("name")).Run(m.session)
-	if err != nil {
-		return nil, err
-	}
-	providers := []*model.Provider{}
-	if err := res.All(&providers); err != nil {
-		return nil, err
-	}
-	return providers, nil
-}
-
-func (m DefaultManager) GetProvider(providerId string) (*model.Provider, error) {
-	res, err := r.Table(tblNameProviders).Filter(map[string]string{"id": providerId}).Run(m.session)
-	if err != nil {
-		return nil, err
-	}
-	if res.IsNil() {
-		return nil, ErrProviderDoesNotExist
-	}
-	var provider *model.Provider
-	if err := res.One(&provider); err != nil {
-		return nil, err
-	}
-	return provider, nil
-}
-
-func (m DefaultManager) CreateProvider(provider *model.Provider) error {
-	var eventType string
-
-	prov, err := m.GetProvider(provider.ID)
-	if err != nil && err != ErrProviderDoesNotExist {
-		return err
-	}
-	if prov != nil {
-		return ErrProviderExists
-	}
-
-	response, err := r.Table(tblNameProviders).Insert(provider).RunWrite(m.session)
-
-	if err != nil {
-		return err
-	}
-	eventType = "add-provider"
-
-	provider.ID = func() string {
-		if len(response.GeneratedKeys) > 0 {
-			return string(response.GeneratedKeys[0])
-		}
-		return ""
-	}()
-
-	m.logEvent(eventType, fmt.Sprintf("id=%s, name=%s", provider.ID, provider.Name), []string{"security"})
-
-	return nil
-}
-
-func (m DefaultManager) UpdateProvider(provider *model.Provider) error {
-	var eventType string
-
-	// check if exists; if so, update
-	prov, err := m.GetProvider(provider.ID)
-	if err != nil && err != ErrProviderDoesNotExist {
-		return err
-	}
-	// update
-
-	if prov != nil {
-		updates := map[string]interface{}{
-			"name":              provider.Name,
-			"availableJobTypes": provider.AvailableJobTypes,
-			"config":            provider.Config,
-			"url":               provider.Url,
-			"providerJobs":      provider.ProviderJobs,
-		}
-
-		if _, err := r.Table(tblNameProviders).Filter(map[string]string{"id": provider.ID}).Update(updates).RunWrite(m.session); err != nil {
-			return err
-		}
-
-		eventType = "update-provider"
-	}
-
-	m.logEvent(eventType, fmt.Sprintf("id=%s, name=%s", provider.ID, provider.Name), []string{"security"})
-
-	return nil
-}
-
-func (m DefaultManager) DeleteProvider(providerId string) error {
-	res, err := r.Table(tblNameProviders).Filter(map[string]string{"id": providerId}).Delete().Run(m.session)
-	if err != nil {
-		return err
-	}
-
-	if res.IsNil() {
-		return ErrProviderDoesNotExist
-	}
-
-	m.logEvent("delete-provider", fmt.Sprintf("id=%s", providerId), []string{"security"})
-
-	return nil
-}
-
-func (m DefaultManager) GetJobsByProviderId(providerId string) ([]*model.ProviderJob, error) {
-	res, err := r.Table(tblNameProviders).Filter(map[string]string{"id": providerId}).Run(m.session)
-	if err != nil {
-		return nil, err
-	}
-	if res.IsNil() {
-		return nil, ErrProviderDoesNotExist
-	}
-	var provider *model.Provider
-	if err := res.One(&provider); err != nil {
-		return nil, err
-	}
-	return provider.ProviderJobs, nil
-}
-
-func (m DefaultManager) AddJobToProviderId(providerId string, job *model.ProviderJob) error {
-
-	var eventType string
-
-	res, err := r.Table(tblNameProviders).Filter(map[string]string{"id": providerId}).Run(m.session)
-	if err != nil {
-		return err
-	}
-	if res.IsNil() {
-		return ErrProviderDoesNotExist
-	}
-	var provider *model.Provider
-	if err := res.One(&provider); err != nil {
-		return err
-	}
-
-	provider.ProviderJobs = append(provider.ProviderJobs, job)
-
-	if _, err := r.Table(tblNameProviders).Filter(map[string]string{"id": provider.ID}).Update(provider).RunWrite(m.session); err != nil {
-		return err
-	}
-	eventType = "add-job-to-provider"
-
-	m.logEvent(eventType, fmt.Sprintf("id=%s, name=%s", provider.ID, provider.Name), []string{"security"})
-
-	return nil
-}
-
-func (m DefaultManager) DeleteAllProviders() error {
-	_, err := r.Table(tblNameProviders).Delete().Run(m.session)
-
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// end methods related to the Image structure
 
 func (m DefaultManager) AddRegistry(registry *model.Registry) error {
